refactor(service): 重构BaseService接口并添加BaseServiceImpl实现在`BaseService`接口中进行了以下更改： - 移除了泛型参数`<ID>`，简化了接口定义。 - 添加了基于MyBatis-Plus的CRUD操作方法，如`save`, `removeById`, `removeByIds`, `updateById`, `saveOrUpdate`, `getById`, `list`, `list(Wrapper<T>)`, `page(PageParam)`, 和 `page(PageParam, Wrapper<T>)`。- 删除了一些不再需要的方法，如`batchSave`和`removeById`（旧版本）。
新增了`BaseServiceImpl`类，实现了`BaseService`接口，并提供了基于MyBatis-Plus的通用实现。该实现类封装了基础的CRUD操作，并支持自定义分页查询。

此外，删除了未使用的`LocalTokenService`接口。

+++ src/main/java/com/community_shop/backend/dto/PageResult.java
@@ -11,15 +11,15 @@ import java.util.List;
 @Data
 public class PageResult<T> {
     // 总记录数
-    private Integer total;
+    private Long total;
     // 总页数
-    private Integer totalPages;
+    private Long totalPages;
     // 当前页数据列表
     private List<T> list;
     // 当前页码
-    private Long pageNum;
+    private Integer pageNum;
     // 每页条数
-    private Long pageSize;
+    private Integer pageSize;
 
     // @Data 不会提供构造函数
     // 添加无参构造函数
